refactor: extract Qcow2Info and Qcow2DevParams into dev/info.rs
Move src/dev.rs to src/dev/mod.rs and split the standalone Qcow2Info
and Qcow2DevParams data types into a new dev/info.rs submodule,
re-exported from mod.rs. Two Qcow2Info helper methods used by the
allocator (rb_slice_entries, is_read_only) are widened to pub(crate)
since they are now accessed across the module boundary. No logic
changes.
